+++ src/selection_sort.h
@@ -16,7 +16,7 @@ namespace runtime {
 }
 
 namespace runcpr {
-    void SelectionSort(int a[], int n, int &comparison) {
+    void SelectionSort(int a[], int n, long long &comparison) {
         for (int i = 0; ++comparison && i < n - 1; ++i) {
             int minIdx = i;
             for (int j = i + 1; ++comparison && j < n; ++j) {
